+++ CHANGELOG.md
@@ -94,6 +94,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [Library] Add chips components ([#407](https://github.com/Orange-OpenSource/ouds-ios/issues/407))
 - [Library] Checkbox picker component ([#587](https://github.com/Orange-OpenSource/ouds-ios/issues/587))
 - [Library] Switch component ([#405](https://github.com/Orange-OpenSource/ouds-ios/issues/405))
 - [Tool] New system to generate, build ad upload documentation ([#628](https://github.com/Orange-OpenSource/ouds-ios/issues/628))